refactor: view에 관련된 컴포넌트 패키지 "view"로 이동
